removed unnecessary casts.
git-svn-id: http://svn.code.sf.net/p/graphiti-editor/code@499 04dbf5ff-8551-0410-9f1f-ce2c5325ca72

+++ editor/src/net/sf/graphiti/ui/editors/GraphitiPalette.java
@@ -189,7 +189,7 @@ public class GraphitiPalette {
 		// adjust width and height
 		double ratio = (double) width / (double) height;
 		width = 16;
-		height = (int) ((double) width / ratio);
+		height = (int) (width / ratio);
 
 		// Creates a new vertex figure
 		Font font = Display.getDefault().getSystemFont();
